spotify-led-controller/requirements.txt
spotipy>=2.23.0
pyserial>=3.5
python-dotenv>=1.0.0
requests>=2.28.0
Pillow>=10.0.0

# Knob mixer dependencies
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
